Refactor time representation from double to long across simulation strategies and related components

+++ SimNextgenApp/Core/IRunContext.cs
@@ -15,7 +15,7 @@ public interface IRunContext
     /// <summary>
     /// Gets the current simulation time in simulation time units.
     /// </summary>
-    double ClockTime { get; }
+    long ClockTime { get; }
 
     /// <summary>
     /// Gets the number of events that have been executed so far in the simulation run.
